Create Variables_00
We have a Bash script that assigns values to variables and displays some information. However, there is a mistake in the script that needs to be fixed. #SPOILER AHEAD# The issue lies in the lines where we are assigning values to the name and age variables. Bash is very particular when it comes to spacing in certain logic.

After fixing the issue, the script should work as expected, correctly displaying the names and ages. Remember to pay attention to Bash's syntax and rules, especially when it comes to variable assignments and spacing.